+++ netty-loom-spring-boot-starter/src/main/java/io/github/azholdaspaev/nettyloomspring/autoconfigure/NettyLoomAutoConfiguration.java
@@ -15,7 +15,6 @@ import io.github.azholdaspaev.nettyloomspring.core.handler.HttpRequestHandler;
 import io.github.azholdaspaev.nettyloomspring.core.pipeline.NettyPipelineStep;
 import io.github.azholdaspaev.nettyloomspring.core.pipeline.NettyPipelineDefinition;
 import io.github.azholdaspaev.nettyloomspring.core.server.NettyIoHandlerFactory;
-import io.github.azholdaspaev.nettyloomspring.core.server.NettyServerChannelInitializer;
 import io.github.azholdaspaev.nettyloomspring.mvc.handler.SpringHttpRequestDispatcher;
 import io.github.azholdaspaev.nettyloomspring.mvc.servlet.DefaultNettyServletContext;
 import io.github.azholdaspaev.nettyloomspring.mvc.servlet.NettyServletContext;
@@ -48,12 +47,12 @@ public class NettyLoomAutoConfiguration {
 
     @Bean
     public NettyWebServerFactory nettyWebServerFactory(NettyIoHandlerFactory nettyIoHandlerFactory,
-                                                       NettyServerChannelInitializer nettyServerChannelInitializer,
+                                                       NettyPipelineDefinition nettyPipelineDefinition,
                                                        HttpConnectionRegistry httpConnectionRegistry,
                                                        NettyServletContext servletContext,
                                                        DispatcherServlet dispatcherServlet,
                                                        NettyLoomProperties properties) {
-        return new NettyWebServerFactory(nettyIoHandlerFactory, nettyServerChannelInitializer,
+        return new NettyWebServerFactory(nettyIoHandlerFactory, nettyPipelineDefinition,
             httpConnectionRegistry, servletContext, dispatcherServlet, properties);
     }
 
@@ -75,12 +74,6 @@ public class NettyLoomAutoConfiguration {
     @Bean
     public HttpConnectionRegistry httpConnectionRegistry() {
         return new HttpConnectionRegistry(new DefaultChannelGroup("netty-loom-channels", GlobalEventExecutor.INSTANCE));
-    }
-
-    @Bean
-    public NettyServerChannelInitializer nettyServerChannelInitializer(NettyPipelineDefinition nettyPipelineDefinition,
-                                                                       HttpConnectionRegistry httpConnectionRegistry) {
-        return new NettyServerChannelInitializer(nettyPipelineDefinition, httpConnectionRegistry);
     }
 
     @Bean